Renamed `utils` to `smelter_state`
