+++ integration_tests/examples/demo/smelter_state/inputs.rs
