+++ integration_tests/examples/demo/smelter_state/inputs/rtp.rs
@@ -10,7 +10,7 @@ use serde_json::json;
 use strum::{Display, EnumIter, IntoEnumIterator};
 use tracing::error;
 
-use crate::utils::{
+use crate::smelter_state::{
     get_free_port,
     inputs::{AudioDecoder, InputHandler, VideoDecoder},
     players::InputPlayerOptions,
