+++ integration_tests/examples/demo/smelter_state/outputs.rs
@@ -4,8 +4,6 @@ use anyhow::Result;
 use serde_json::json;
 use strum::{Display, EnumIter};
 
-use crate::utils::{inputs::InputHandler, TransportProtocol};
-
 pub mod rtmp;
 pub mod rtp;
