+++ integration_tests/examples/demo/smelter_state/outputs/rtmp.rs
